osbuild-worker: parallelise jobs based on kind
Run jobs with similar latency requirements, CPU and IO performance in the same
loop. This avoids low-latency jobs being slowed down by long-running ones.

On busy workers this should in particular improve koji jobs as the koji-init and
koji-finalie no longer need to wait for unrelated osbuild jobs to finish.

In the future we extend this to run several loops of mostly idle jobs (waiting on
external dependencies) in parallel.