jobImplementation/Run: support interruption
Pass a context to Run, which is cancelled on SIGTERM. Each job
type is responsible for handling this separately. For now none
of them do.

If a run has been cancelled, the job is not updated with a result.
This allows the job to be rescheduled by composer and avoids
misleading partial results.

Note that composer will not know that a job has been interrupted,
and only when the heartbeat expires will it reschedule the job.
We do this twice before giving up, which is not ideal, as
interrupted workers should be part of normal operations, so we
should ideally retry more times before giving up. This will be
addressed in a future commit.