osbuild-worker: introduce global async instance protection
Introduce a global async instance protector.

The instance protector attempts to ask the cloud providers to keep the VM running for the duration of our jobs, and replaces the existing inline protection.

This remains best effort only, and the cloud provider can ignore the request and
the requests are known to be racy. That means we still need to handle VMs
being shut down mid job.

The change from the existing logic:
 - no longer blocking, but happens in the background
 - multiple jobs can use it simultaneously
 - all job types use it in the same way
 - grace periods to avoid busy-loops
 - collation of redundant requests

See inline comments for more details.