tools: silence version comparison in get_build_info()
The get_build_info() function uses nvrGreaterOrEqual() to determine how
to parse the output from composer-cli.  The function prints the result
of the comparison but is often used in a loop when waiting for a compose
to finish, which makes CI logs very annoying.

Silence the version comparison output when called form get_build_info().