worker/client: move job runner from cmd/
Move the run loop implementation from cmd/osbuild-worker and make
it part of the internal library instead.

Additionally, support a context for all calls, allowing long-running
calls like requesting a new job or uploading an artifact to be
cancelled.

In the worker package, make methods internal when possible, and
drop the Job interface, in favor of an explicit type.