distro/rhel8: rename raw image in simplified installer
Rename the raw image embedded in the simplified-installer ISO to
image.raw.xz for RHEL 8.  This is the same name that's used for the raw
image when built on its own (as edge-raw-image).
The change is made for consistency with RHEL 9 and Fedora as well as for
consistency with the edge-raw-image.